+++ src/types/wallet.type.ts
@@ -1,0 +1,3 @@
+export interface WalletAddressType{
+    readonly walletAddress : string | undefined
+}
